Bumped seek whence up to uint32_t
I can't think of a reason this should be uint8_t. Bumping it up to
uint32_t matches the type used for other flags (even though whence is
arguably not flags in a strict sense).

No code changes.

+++ lfs3.c
@@ -14282,7 +14282,7 @@ failed:;
 // other file operations
 
 lfs3_soff_t lfs3_file_seek(lfs3_t *lfs3, lfs3_file_t *file,
-        lfs3_soff_t off, uint8_t whence) {
+        lfs3_soff_t off, uint32_t whence) {
     LFS3_ASSERT(lfs3_handle_isopen(lfs3, &file->b.h));
 
     // TODO check for out-of-range?
